Improve addition of Java to Docker image
Previously, the Docker file added a bunch of utilities and downloaded the version of Java directly in the image that was eventually created.  This left a bunch of unnecessary and potentially vulnerable packages on the image that was used in production.  This change makes the build a multi-stage build and ensures that the network utilities required for downloading only exist on a disposed stage.

In addition to the change to a multi-stage build, this change also swaps from the Pivotal Distribution of OpenJDK to AdoptOpenJDK as part of our commitment to move to an industry standard distribution.  It also swaps from Java 8 to Java 11.  Trust me, you'll be fine.