Improve Addition of Java to Docker Image
Previously, the Docker file added a bunch of utilities and downloaded the
version of Java directly in the image that was eventually created.  This left
a bunch of unnecessary and potentially vulnerable packages on the image that
was used in production.  This change makes the build a multi-stage build and
ensures that the network utilities required for downloading only exist on a
disposed stage.

In addition to the change to a multi-stage build, this change also swaps from
the Pivotal Distribution of OpenJDK to BellSoft Liberica.

Signed-off-by: Ben Hale <bhale@vmware.com>